Processor::Utilization is float

+++ include/processor.h
@@ -6,7 +6,7 @@
 
 class Processor {
  public:
-  double Utilization();  // TODO: See src/processor.cpp
+  float Utilization();  // TODO: See src/processor.cpp
 
   // TODO: Declare any necessary private members
  private:
