+++ src/processor.cpp
@@ -2,7 +2,7 @@
 #include "processor.h"
 #include "linux_parser.h"
 
-double Processor::Utilization() {
+float Processor::Utilization() {
     long const prev_idle = LinuxParser::IdleJiffies();
     long const prev_nonidle = LinuxParser::ActiveJiffies();
     std::this_thread::sleep_for(std::chrono::milliseconds(100));
@@ -10,5 +10,5 @@ double Processor::Utilization() {
     long nonidle = LinuxParser::ActiveJiffies();
     idle = idle - prev_idle;
     nonidle = nonidle - prev_nonidle;
-    return (double) nonidle / (nonidle + idle);
+    return (float) nonidle / (nonidle + idle);
 }
